+++ lib/client/webdav_client.dart
@@ -1,6 +1,8 @@
 import 'dart:convert' show utf8;
+import 'dart:io';
 
 import 'package:dio/dio.dart';
+import 'package:dio/io.dart';
 import 'package:ntodotxt/main.dart' show log;
 import 'package:webdav_client/webdav_client.dart' as webdav;
 
@@ -14,38 +16,100 @@ class WebDAVClientException implements Exception {
 }
 
 class WebDAVClient {
-  late final webdav.Client client;
+  webdav.Client? _client; // Singleton pattern
+
+  late final String scheme;
+  late final String host;
+  late final int? port;
+  late final String path;
+  final String username;
+  final String password;
+  final bool acceptUntrustedCert;
+
+  static const int idleTimeout = 15000;
+  static const int connectionTimeout = 15000;
+  static const int sendTimeout = 15000;
+  static const int receiveTimeout = 15000;
 
   WebDAVClient({
-    String schema = 'http',
-    required String host,
-    int? port,
-    required String baseUrl,
-    required String username,
-    required String password,
+    required String server,
+    required String path,
+    required this.username,
+    required this.password,
+    this.acceptUntrustedCert = false,
   }) {
-    if (baseUrl.endsWith('/')) {
-      baseUrl = baseUrl.substring(0, baseUrl.length - 1);
+    final RegExp exp = RegExp(
+        r'(?<scheme>^(http|https)):\/\/(?<host>[a-zA-Z0-9.-]+)(:(?<port>\d+)){0,1}$');
+    final RegExpMatch? match = exp.firstMatch(server);
+    if (match != null) {
+      scheme = match.namedGroup('scheme')!;
+      host = match.namedGroup('host')!;
+      port = match.namedGroup('port') != null
+          ? int.parse(match.namedGroup('port')!)
+          : null;
+    } else {
+      throw const FormatException('Invalid server format');
     }
-    client = webdav.newClient(
-      Uri(
-        scheme: schema,
+
+    if (path == '/') {
+      this.path = path;
+    } else {
+      if (path.endsWith('/')) {
+        this.path = path.substring(0, path.length - 1);
+      } else {
+        this.path = path;
+      }
+    }
+  }
+
+  Future<webdav.Client> get connection async {
+    if (_client != null) {
+      return _client!;
+    } else {
+      _client = await _open();
+      return _client!;
+    }
+  }
+
+  Future<webdav.Client> _open() async {
+    // Handle untrusted certificates.
+    webdav.WdDio dio = webdav.WdDio(debug: false);
+    dio.httpClientAdapter = IOHttpClientAdapter(
+      createHttpClient: () {
+        // Don't trust any certificate just because their root cert is trusted.
+        final HttpClient httpClient =
+            HttpClient(context: SecurityContext(withTrustedRoots: false));
+        httpClient.idleTimeout = const Duration(
+          milliseconds: idleTimeout,
+        );
+        httpClient.connectionTimeout = const Duration(
+          milliseconds: connectionTimeout,
+        );
+        // You can test the intermediate / root cert here. We just ignore it.
+        httpClient.badCertificateCallback =
+            (cert, host, port) => acceptUntrustedCert;
+        return httpClient;
+      },
+      validateCertificate: (cert, host, port) => acceptUntrustedCert,
+    );
+
+    webdav.Client client = webdav.Client(
+      uri: Uri(
+        scheme: scheme,
         host: host,
         port: port,
-        path: baseUrl,
+        path: path,
       ).toString(),
-      user: username,
-      password: password,
+      c: dio,
+      auth: webdav.BasicAuth(user: username, pwd: password),
       debug: false,
     );
-    // Set the public request headers
     client.setHeaders({'accept-charset': 'utf-8'});
-    // Set the connection server timeout time in milliseconds.
-    client.setConnectTimeout(15000);
-    // Set send data timeout time in milliseconds.
-    client.setSendTimeout(15000);
-    // Set transfer data time in milliseconds.
-    client.setReceiveTimeout(15000);
+    client.setConnectTimeout(connectionTimeout);
+    client.setSendTimeout(sendTimeout);
+    client.setReceiveTimeout(receiveTimeout);
+
+    return client;
   }
 
   (String, String) _handleDioError(DioException error) {
@@ -53,82 +117,91 @@ class WebDAVClient {
       case DioExceptionType.connectionTimeout:
       case DioExceptionType.sendTimeout:
       case DioExceptionType.receiveTimeout:
-        return ('Timeout', 'Timeout occurred while sending or receiving');
+        return (
+          'TIMEOUT',
+          'Timeout occurred while sending or receiving',
+        );
       case DioExceptionType.badResponse:
         final statusCode = error.response?.statusCode;
         if (statusCode != null) {
           switch (statusCode) {
             case 400:
               return (
-                'Bad Request',
+                'BAD REQUEST',
                 'Something went wrong',
               );
             case 401:
               return (
-                'Unauthorized',
+                'UNAUTHORIZED',
                 'It seems that the credentials are incorrect',
               );
             case 403:
               return (
-                'Forbidden',
+                'FORBIDDEN',
                 'The request was rejected by the server',
               );
             case 404:
               return (
-                'Not Found',
+                'NOT FOUND',
                 'The requested resource could not be found',
               );
             case 405:
               return (
-                'Method Not Allowed',
+                'METHOD NOT ALLOWED',
                 'The request method is not supported for the requested resource',
               );
             case 409:
               return (
-                'Conflict',
+                'CONFLICT',
                 'The request could not be processed because of conflict in the current state of the resource',
               );
             case 500:
               return (
-                'Internal Server Error',
-                'An unexpected error was encountered',
+                'INTERNAL SERVER ERROR',
+                'An unexpected error was encountered on server side',
+              );
+            default:
+              return (
+                'INTERNAL SERVER ERROR ($statusCode)',
+                'An unexpected error was encountered on server side',
               );
           }
         }
         break;
       case DioExceptionType.cancel:
         break;
-      case DioExceptionType.unknown:
-        return (
-          'Unknown Error',
-          'Server cannot be reached',
-        );
       case DioExceptionType.badCertificate:
         return (
-          'Internal Server Error',
+          'INVALID SSL CERTIFICATE',
           'The certificate is invalid',
         );
       case DioExceptionType.connectionError:
         return (
-          'Connection Error',
+          'CONNECTION ERROR',
           'Server cannot be reached',
+        );
+      case DioExceptionType.unknown:
+        return (
+          'UNKNOWN ERROR',
+          'Possible cause may be a connection problem or an invalid SSL certificate',
         );
       default:
         return (
-          'Unknown Error',
-          'Something went wrong',
+          'UNKNOWN ERROR',
+          'Possible cause may be a connection problem or an invalid SSL certificate',
         );
     }
 
     return (
-      'Unknown Error',
-      'Something went wrong',
+      'UNKNOWN ERROR',
+      'Something unexpected went wrong',
     );
   }
 
   Future<void> ping() async {
     try {
       log.fine('Ping');
+      webdav.Client client = await connection;
       await client.ping();
     } on DioException catch (e) {
       log.severe(e);
@@ -193,6 +266,7 @@ class WebDAVClient {
   Future<List<webdav.File>> listFiles({
     String path = '',
   }) async {
+    webdav.Client client = await connection;
     try {
       log.fine('List files and directories of ${path.isEmpty ? "/" : path}');
       return await client.readDir(path);
@@ -205,6 +279,7 @@ class WebDAVClient {
   }
 
   Future<void> create(String filename) async {
+    webdav.Client client = await connection;
     try {
       if (filename.startsWith('/')) {
         filename = filename.substring(1);
@@ -226,6 +301,7 @@ class WebDAVClient {
 
   Future<void> mkdir(
       {required String directory, bool recursive = false}) async {
+    webdav.Client client = await connection;
     if (directory.startsWith('/')) {
       directory = directory.substring(1);
     }
@@ -250,6 +326,7 @@ class WebDAVClient {
   Future<String> download({
     required String filename,
   }) async {
+    webdav.Client client = await connection;
     try {
       log.fine('Download content of file $filename');
       List<int> content = await client.read(filename);
@@ -266,6 +343,7 @@ class WebDAVClient {
     required String filename,
     required String content,
   }) async {
+    webdav.Client client = await connection;
     try {
       log.fine('Upload content to file $filename');
       await client.write(filename, utf8.encode(content));