+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Checkbox to accept untrusted SSL certificates #72
+
 ### Changed
 
 - Removes the save button and adds a save/discard dialog instead when leaving the todo/filter page #61